requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
boto3==1.16.38
botocore==1.19.38
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
chowkidar-graphene==0.1.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.4
django-ses==1.0.3
django-storages==1.11
environs==9.2.0
future==0.18.2
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
huey==2.3.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
marshmallow==3.9.1
promise==2.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
Rx==1.6.1
s3transfer==0.3.3
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.26.2
uuid==1.30
zope.interface==5.2.0
